move loginSlice to app
