+++ src/app/store.ts
@@ -1,6 +1,6 @@
 import { configureStore, ThunkAction, Action } from '@reduxjs/toolkit';
 import counterReducer from '../features/counter/counterSlice';
-import loginReducer from '../pages/Landing/loginSlice';
+import loginReducer from './loginSlice';
 
 export const store = configureStore({
   reducer:{
