+++ src/pages/GraphView/GraphView.tsx
@@ -1,5 +1,5 @@
 import Graph from 'react-graph-vis';
-import { useState } from 'react';
+import React, { useState } from 'react';
 import styled from 'styled-components/macro';
 
 const Wrapper = styled.div`
